Change election template data
The idea is to create election types that wraps common settings to
algorithms. E.g. an election of type list election would also need
settings, so that the voter screens look right for the voter, and that
the right settings are used for counting. This way, it should be easier
to reuse code.