Move "Blank vote" above "Review ballot"
In the interface for mobile phones (and the like), the button for submitting a
blank vote kind of disappears beneath the button for reviewing a ballot.

It can be argued that this hides the option of submitting a blank vote,
therefore we move "Blank vote" above "Review ballot", so it is clear for
everyone that this is a choice.