Request viewer first
A problem used to occur on first login by a user. Once the user got to
the front page, multiple queries where fired simultaneously. This
triggered multiple person objects being created for that user in
parallel, which would lead to duplication errors in the database.
All of this is avoided if the viewer query is fired first.

Issue: EVALG-620